+++ R/Algorithm.R
@@ -43,6 +43,7 @@ addAlgorithm = function(name, fun = NULL, reg = getDefaultRegistry())  {
     assert(checkFunction(fun, args = c("job", "data", "instance")), checkFunction(fun, args = "..."))
   }
 
+  info("Adding algorithm '%s'", name)
   algo = setClasses(list(fun = fun, name = name), "Algorithm")
   writeRDS(algo, file = getAlgorithmURI(reg, name))
   reg$defs$algorithm = addlevel(reg$defs$algorithm, name)
@@ -63,7 +64,7 @@ removeAlgorithms = function(name, reg = getDefaultRegistry()) {
     job.ids = filter(def.ids, reg$status, "job.id")
 
     info("Removing Algorithm '%s' and %i corresponding jobs ...", nn, nrow(job.ids))
-    file.remove(getAlgorithmURI(reg, nn))
+    file.remove.safely(getAlgorithmURI(reg, nn))
     reg$defs = reg$defs[!def.ids]
     reg$status = reg$status[!job.ids]
     reg$defs$algorithm = rmlevel(reg$defs$algorithm, nn)
@@ -78,8 +79,4 @@ removeAlgorithms = function(name, reg = getDefaultRegistry()) {
 getAlgorithmIds = function(reg = getDefaultRegistry()) {
   assertExperimentRegistry(reg)
   levels(reg$defs$algorithm)
-}
-
-getAlgorithmURI = function(reg, name) {
-  file.path(reg$file.dir, "algorithms", mangle(name))
 }