+++ R/clusterFunctionsSGE.R
@@ -25,7 +25,8 @@
 #' @return [\code{\link{ClusterFunctions}}].
 #' @family ClusterFunctions
 #' @export
-makeClusterFunctionsSGE = function(template = findTemplateFile("sge"), scheduler.latency = 1, fs.latency = 65) { # nocov start
+makeClusterFunctionsSGE = function(template = "sge", scheduler.latency = 1, fs.latency = 65) { # nocov start
+  template = findTemplateFile(template)
   template = cfReadBrewTemplate(template)
 
   submitJob = function(reg, jc) {
@@ -33,7 +34,7 @@ makeClusterFunctionsSGE = function(template = findTemplateFile("sge"), scheduler
     assertClass(jc, "JobCollection")
 
     outfile = cfBrewTemplate(reg, template, jc)
-    res = runOSCommand("qsub", outfile)
+    res = runOSCommand("qsub", shQuote(outfile))
 
     if (res$exit.code > 0L) {
       cfHandleUnknownSubmitError("qsub", res$exit.code, res$output)