+++ tests/testthat/test_findConfFile.R
@@ -4,8 +4,7 @@ test_that("findConfFile", {
   d = tempdir()
   fn = file.path(d, "batchtools.conf.R")
   writeLines("hello", fn)
-  prev = getwd()
-  setwd(d)
-  expect_equal(findConfFile(), npath(fn))
-  setwd(prev)
+  withr::with_dir(d,
+    expect_equal(findConfFile(), normalizePath(fn))
+  )
 })